Added debugging and a cron.sh file
I had to add all the debug to check why the database was left locked
when the cron was running. Turns out Mastodon's API suddently started
to take forever to respond (like almost 2 minutes). So yeah, added a
bunch of _debug() with times to see how things run if needed.

Also the cron.sh was made so I can have a PID file that "expires" after
30 minutes, just in case. That way we don't run the script again while
it's still waiting for Mastodon or any other 3rd party thing to reply.